# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AKB1 Delivery Command Center",
    "version": "0.3.0-backend"
  },
  "paths": {
...
    "/api/v1/admin/tier-config/{tier_number}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Patch Tier Config",
        "description": "PATCH a tier_config row.\n\nPO-only. Writes the full before / after snapshot to audit_trail_entries\nin the same transaction; if the audit write fails, the UPDATE rolls back.",
        "operationId": "patch_tier_config_api_v1_admin_tier_config__tier_number__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tier_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tier Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TierConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tier-config": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Tier Config List",
        "description": "Read every escalation tier config. PO-only. Reads not audited.",
        "operationId": "get_tier_config_list_api_v1_admin_tier_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierConfigList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/threshold-register": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Threshold Register List",
        "description": "Read every threshold calibration register row. PO-only. Reads not audited.",
        "operationId": "get_threshold_register_list_api_v1_admin_threshold_register_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThresholdRegisterList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/threshold-register/{metric_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Patch Threshold Register",
        "description": "PATCH a threshold_calibration_register row.\n\nPO-only. Writes the full before / after snapshot to audit_trail_entries\nin the same transaction; if the audit write fails, the UPDATE rolls back.",
        "operationId": "patch_threshold_register_api_v1_admin_threshold_register__metric_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "metric_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Metric Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThresholdRegisterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThresholdRegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ThresholdRegisterList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ThresholdRegisterResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ThresholdRegisterList",
        "description": "GET /api/v1/admin/threshold-register response."
      },
      "ThresholdRegisterResponse": {
        "properties": {
          "metric_id": {
            "type": "string",
            "title": "Metric Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "green_threshold": {
            "type": "string",
            "title": "Green Threshold"
          },
          "amber_threshold": {
            "type": "string",
            "title": "Amber Threshold"
          },
          "red_threshold": {
            "type": "string",
            "title": "Red Threshold"
          },
          "range_lower": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range Lower"
          },
          "range_upper": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range Upper"
          },
          "rationale_text": {
            "type": "string",
            "title": "Rationale Text"
          },
          "last_calibrated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Calibrated At"
          },
          "last_calibrated_by": {
            "type": "string",
            "format": "uuid",
            "title": "Last Calibrated By"
          },
          "owning_role": {
            "type": "string",
            "title": "Owning Role"
          }
        },
        "type": "object",
        "required": [
          "metric_id",
          "display_name",
          "direction",
          "green_threshold",
          "amber_threshold",
          "red_threshold",
          "range_lower",
          "range_upper",
          "rationale_text",
          "last_calibrated_at",
          "last_calibrated_by",
          "owning_role"
        ],
        "title": "ThresholdRegisterResponse"
      },
      "ThresholdRegisterUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(HigherIsBetter|LowerIsBetter|RangeIsBetter)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "green_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Green Threshold"
          },
          "amber_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amber Threshold"
          },
          "red_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Red Threshold"
          },
          "range_lower": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range Lower"
          },
          "range_upper": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range Upper"
          },
          "rationale_text": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale Text"
          },
          "owning_role": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(PortfolioOwner|FinanceLead|ProgrammeManager)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owning Role"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ThresholdRegisterUpdate",
        "description": "Body for PATCH /api/v1/admin/threshold-register/{metric_id}.\n\nAll fields optional; at least one must be provided. extra='forbid'\nrejects unknown keys. Setting range_lower / range_upper to null in\nthe request body is treated as \"no change\" in this slice; emptying\nthose fields requires a separate slice when needed (see slice 2 hand-back).\n\nmetric_id is the URL path param and not editable. last_calibrated_at\nand last_calibrated_by are system-set on every PATCH."
      },
      "TierConfigList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TierConfigResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TierConfigList",
        "description": "GET /api/v1/admin/tier-config response. Wrapped list for forward\nextensibility (pagination, total_count when row counts grow)."
      },
      "TierConfigResponse": {
        "properties": {
          "tier_number": {
            "type": "integer",
            "title": "Tier Number"
          },
          "default_label": {
            "type": "string",
            "title": "Default Label"
          },
          "display_label": {
            "type": "string",
            "title": "Display Label"
          },
          "description_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Text"
          },
          "sla_hint_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Hint Hours"
          },
          "role_mapping": {
            "type": "object",
            "title": "Role Mapping"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "last_edited_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Edited At"
          },
          "last_edited_by_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Last Edited By User Id"
          }
        },
        "type": "object",
        "required": [
          "tier_number",
          "default_label",
          "display_label",
          "description_text",
          "sla_hint_hours",
          "role_mapping",
          "active",
          "last_edited_at",
          "last_edited_by_user_id"
        ],
        "title": "TierConfigResponse"
      },
      "TierConfigUpdate": {
        "properties": {
          "display_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Label"
          },
          "description_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Text"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TierConfigUpdate",
        "description": "Body for PATCH /api/v1/admin/tier-config/{tier_number}.\n\nAll fields optional; at least one must be provided. extra='forbid'\nrejects unknown keys so contract drift is caught at the boundary."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}